+++ src/classes/game-settings.ts
@@ -379,6 +379,7 @@ export class GameSettings {
                     if(currentDate.year !== newDate.year || currentDate.month !== newDate.month || currentDate.day !== newDate.day || currentDate.seconds !== newDate.seconds){
                         await (<Game>game).settings.set(ModuleName, SettingNames.CurrentDate, newDate);
                         if(emitHook && SimpleCalendar.instance){
+                            await GameSockets.emit(<SimpleCalendarSocket.Data>{type: SocketTypes.emitHook, data: <SimpleCalendarSocket.SimpleCalendarEmitHook>{hook: SimpleCalendarHooks.DateTimeChange}});
                             Hook.emit(SimpleCalendarHooks.DateTimeChange);
                         }
                         await GameSockets.emit(<SimpleCalendarSocket.Data>{ type: SocketTypes.noteReminders, data: { justTimeChange: currentDate.seconds !== newDate.seconds && currentDate.day === newDate.day }});